requirements.txt
absl-py==0.3.0
astor==0.7.1
certifi==2020.6.20
chardet==3.0.4
click==6.7
Flask==1.0.2
gast==0.2.0
grpcio==1.14.0
gunicorn==19.9.0
idna==2.10
importlib-metadata==1.7.0
itsdangerous==0.24
Jinja2==2.10
joblib==0.16.0
jsonpickle==1.4.1
Markdown==3.2.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
numpy==1.15.0
opencv-python==3.4.2.17
protobuf==3.6.0
python-sql==1.1.0
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.1
six==1.11.0
sklearn==0.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3==1.25.9
Werkzeug==0.16.1
zipp==3.1.0
